Awais Admin part works
The sql query returns an array, we pul the 2nd index from thearray and use the getAdmin metohod to get a boolean dataype. Thenm a check is done to see if it is is true or false. If it is is true then an admin has logged, if it is false then a normal user has logged in. Hammad needs to redirect the cheks to a frint end.